backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
aiohttp==3.10.10
elasticsearch==8.15.1
arrow==1.3.0
fire==0.7.0
html2text==2024.2.26
joblib==1.4.2
minio==7.2.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pymilvus==2.4.9
PyMuPDF==1.24.13
requests==2.32.3
tqdm==4.67.0
PyMySQL==1.1.0
SQLAlchemy==2.0.36
cryptography==44.0.0
python-jose==3.3.0
